+++ CHANGELOG.md
@@ -2,12 +2,14 @@
 
 ### Updated
 
-* genre: exclude label name, unless it maps to a valid MusicBrainz genre
+* `genre`: exclude label name, unless it maps to a valid MusicBrainz genre
 
 ### Fixed
 
-* album: When album name contains **Album (Label something)**, the Label is kept in place
+* `album`: When album name contains **Album (Label something)**, the Label is kept in place
   since it immediately follows an opening parenthesis.
+
+* `title`: handle an edge case where one of the track names contains a utf-8 equivalent of a dash
 
 
 ## [0.15.0] 2022-05-16